+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Enhancements
 
+- GLS-257 - rollout Market Guide v1 with data pipeline stats
 - CHORE - Update node requirement to v16
 
 ## [2.9.0](https://github.com/uktrade/great-cms/releases/tag/2.9.0)